refactor: Issue Table row 별도 컴포넌트 분리
- 공통 스타일 믹스인
