+++ src/components/molecules/issue-table-row/index.module.scss
@@ -1,0 +1,5 @@
+@import "../../../styles";
+
+.tr {
+  @include issue-table-tr;
+}
