+++ src/styles/_mixins.scss
@@ -8,3 +8,37 @@
     opacity: 0.7;
   }
 }
+
+@mixin issue-table-tr {
+  padding-left: 1rem;
+  border: 1px solid lightgray;
+  display: flex;
+  align-items: center;
+  min-height: 4rem;
+
+  td {
+    display: flex;
+    align-items: center;
+    height: 100%;
+    &:first-child {
+      flex: 1;
+    }
+
+    &:nth-child(2) {
+      flex: 4;
+    }
+
+    &:last-child {
+      justify-content: flex-end;
+      padding-right: 1rem;
+      flex: 1;
+      div {
+        button {
+          &:first-child {
+            margin-right: 0.5rem;
+          }
+        }
+      }
+    }
+  }
+}
